Changed to powers of three
The efficient.py alg works perfectly fine without any holes if you use
powers of three instead of powers of two. Further, if you check to make
sure the strings are the same length, it is impossible to do a differing
string length crack on it. These have been implemented.

Errors the script would output were valueerrors before, which would
cause a traceback to show too. Now, it prints to stderr and exits with a
code instead of just using valueErrors.