Added version number checker
Added version number checking code to the git repo.
This code was created to solve another algorithmic problem, and
previously was floating outside the repo. Since it is an algo problem
from the CS club, it does belong in this repository.

Also moved .gitignore to the project root where it belongs.